# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "EcoTrack API",
    "version": "1.0.0",
    "description": "API per il sistema EcoTrack"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/api/tipoRaccolta": {
      "get": {
        "summary": "Recupera tutti i tipi di raccolta",
        "description": "Ottiene la lista di tutti i tipi di raccolta disponibili",
        "tags": [
          "Tipo di Raccolta"
        ],
        "responses": {
          "200": {
            "description": "Lista di tipi di raccolta ottenuta con successo"
          },
          "500": {
            "description": "Errore del server"
          }
        }
      },
      "post": {
        "summary": "Crea un nuovo tipo di raccolta",
        "description": "Aggiunge un nuovo tipo di raccolta al sistema",
        "tags": [
          "Tipo di Raccolta"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "tipo": {
                    "type": "string",
                    "description": "Nome del tipo di raccolta (es. \"Carta\", \"Vetro\", \"Plastica\")"
                  }
                },
                "required": [
                  "tipo"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tipo di raccolta creato con successo"
          },
          "400": {
            "description": "Dati non validi"
          },
          "500": {
            "description": "Errore del server"
          }
        }
      }
    },
...
  },
...
}